refactor(price-calculation-service): move constants into utilities & configs
