+++ price-calculation-service/src/main/java/com/cabaggregator/pricecalculationservice/config/GeoGridConfig.java
@@ -4,6 +4,6 @@ import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-public final class GeoGridParams {
+public final class GeoGridConfig {
     public static final double CELL_SIZE = 0.01;
 }
