+++ price-calculation-service/src/main/java/com/cabaggregator/pricecalculationservice/config/cache/CacheTtlConfig.java
@@ -1,4 +1,4 @@
-package com.cabaggregator.pricecalculationservice.config;
+package com.cabaggregator.pricecalculationservice.config.cache;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
