+++ price-calculation-service/src/main/java/com/cabaggregator/pricecalculationservice/config/redis/RedisConfig.java
@@ -1,4 +1,4 @@
-package com.cabaggregator.pricecalculationservice.config;
+package com.cabaggregator.pricecalculationservice.config.redis;
 
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import com.fasterxml.jackson.core.JsonGenerator;
