+++ price-calculation-service/src/main/java/com/cabaggregator/pricecalculationservice/core/constant/StringTemplates.java
@@ -1,0 +1,9 @@
+package com.cabaggregator.pricecalculationservice.core.constant;
+
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public final class StringTemplates {
+    public static final String GRID_CELL = "%.2f,%.2f";
+}
