+++ price-calculation-service/src/main/java/com/cabaggregator/pricecalculationservice/service/impl/GeoGridServiceImpl.java
@@ -1,11 +1,12 @@
 package com.cabaggregator.pricecalculationservice.service.impl;
 
+import com.cabaggregator.pricecalculationservice.core.constant.StringTemplates;
 import com.cabaggregator.pricecalculationservice.service.GeoGridService;
 import org.springframework.stereotype.Service;
 
 import java.util.Locale;
 
-import static com.cabaggregator.pricecalculationservice.config.GeoGridParams.CELL_SIZE;
+import static com.cabaggregator.pricecalculationservice.config.GeoGridConfig.CELL_SIZE;
 
 @Service
 public class GeoGridServiceImpl implements GeoGridService {
@@ -19,6 +20,6 @@ public class GeoGridServiceImpl implements GeoGridService {
         double roundedLat = Math.floor(latitude / CELL_SIZE) * CELL_SIZE;
         double roundedLon = Math.floor(longitude / CELL_SIZE) * CELL_SIZE;
 
-        return String.format(Locale.US,"%.2f,%.2f", roundedLat, roundedLon);
+        return String.format(Locale.US, StringTemplates.GRID_CELL, roundedLat, roundedLon);
     }
 }
